fIX/Changed Id return for user for Long type

+++ app/src/main/java/com/example/geyugoapp/database/dtos/UserDbDto.kt
@@ -1,12 +1,11 @@
 package com.example.geyugoapp.database.dtos
 
 import androidx.room.Entity
-import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "users")
 data class UserDbDto(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val name: String,
     val notificationTime: Long,
     val notificationsActive: Boolean
